pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "clawcity"
version = "0.1.0"
description = "Claw City pipeline for AI generated episodes"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyyaml",
    "openai",
    "replicate",
    "edge-tts",
    "pydub",
    "requests",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
    "black>=24.0.0",
    "flake8>=7.0.0",
]

[project.scripts]
clawcity = "clawcity.cli.main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "C4", "SIM"]
ignore = ["E203", "E501", "E402", "F401", "UP009"]
exclude = [".git", "__pycache__", "build", "dist", "venv", ".venv"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"tests/*" = ["S101"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
ignore_missing_imports = true
namespace_packages = true
explicit_package_bases = true
